:sparkles: add top contributors
