--- src/components/People/People.css
@@ -1,3 +1,0 @@
-.people {
-  max-height: 100vh;
-}
