--- src/components/People/People.js
@@ -1,15 +1,0 @@
-import React from "react";
-
-import "./People.css";
-
-const People = () => {
-  return (
-    <article className="people">
-      <div className="people__title">
-        <h2>People and contributors</h2>
-      </div>
-    </article>
-  );
-};
-
-export default People;
