+++ src/components/Stats/People/index.js
