+++ src/components/Stats/Stats.css
@@ -1,0 +1,4 @@
+.stats {
+  display: flex;
+  width: 100%;
+}
