+++ src/components/Stats/TopProjects/Project/Project.css
@@ -7,6 +7,7 @@
   align-items: center;
   background: #EBECED;
   border-bottom: 1px solid #FFF;
+  width: 100%;
 }
 
 .project__title {
