+++ src/components/Stats/index.js
@@ -1,0 +1,1 @@
+export { default  as Stats } from './Stats';
