+++ src/pages/index.js
@@ -3,7 +3,7 @@ import React from "react"
 import Layout from "../components/layout"
 import SEO from "../components/seo"
 import { Metrics } from "../components/Metrics"
-import { TopProjects } from "../components/TopProjects"
+import { Stats } from "../components/Stats"
 import { NewProjects } from "../components/NewProjects"
 import { Resources } from "../components/Resources"
 
@@ -11,7 +11,7 @@ const IndexPage = () => (
   <Layout>
     <SEO title="Home" keywords={[`gatsby`, `application`, `react`]} />
     <Metrics />
-    <TopProjects />
+    <Stats />
     <NewProjects />
     <Resources />
   </Layout>
